Fix signed / unsigned warning in ParseHelper.cpp. Review URL: https://codereview.appspot.com/7330046
git-svn-id: https://angleproject.googlecode.com/svn/branches/dx11proto@1962 736b8ea6-26fd-11df-bfd4-992fa37f6226

+++ src/compiler/ParseHelper.cpp
@@ -512,7 +512,7 @@ bool TParseContext::constructorErrorCheck(int line, TIntermNode* node, TFunction
     if (constType)
         type->setQualifier(EvqConst);
 
-    if (type->isArray() && type->getArraySize() != function.getParamCount()) {
+    if (type->isArray() && static_cast<size_t>(type->getArraySize()) != function.getParamCount()) {
         error(line, "array constructor needs one argument per array element", "constructor");
         return true;
     }
